requirements.txt
google-generativeai>=0.8.0
selenium>=4.18.0
requests>=2.31.0
